LayoutTests/http/tests/security/xss-DENIED-assign-location-href-javascript-expected.txt

CONSOLE MESSAGE: line 13: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.


--------
Frame: '<!--frame1-->'
--------
You can see the contents of this file.
